+++ slipstream-lib/src/main/java/com/zodl/slipstream/internal/spend/SlipstreamSpendService.kt
@@ -5,6 +5,7 @@ package com.zodl.slipstream.internal.spend
 import cash.z.ecc.android.sdk.exception.PcztException
 import cash.z.ecc.android.sdk.exception.TransactionEncoderException
 import cash.z.ecc.android.sdk.internal.Backend
+import cash.z.ecc.android.sdk.internal.ext.clearContents
 import cash.z.ecc.android.sdk.internal.ext.requireSingleStepForPczt
 import cash.z.ecc.android.sdk.internal.ext.toProposalException
 import cash.z.ecc.android.sdk.internal.transaction.submitTransaction
@@ -125,7 +126,13 @@ internal class SlipstreamSpendService(
     ): Flow<TransactionSubmitResult> =
         flow {
             ensureSaplingParams()
-            val txIds = backend.createProposedTransactions(proposal.toUnsafe(), usk.copyBytes())
+            val uskBytes = usk.copyBytes()
+            val txIds =
+                try {
+                    backend.createProposedTransactions(proposal.toUnsafe(), uskBytes)
+                } finally {
+                    uskBytes.clearContents()
+                }
             engine.notifyTxChange()
             for (txId in txIds) {
                 val txIdBytes = FirstClassByteArray(txId)